Preparing to run keypoint training

+++ references/detection/coco_utils.py
@@ -59,13 +59,14 @@ class ConvertCocoPolysToMask(object):
         anno = [obj for obj in anno if obj['iscrowd'] == 0]
 
         boxes = [obj["bbox"] for obj in anno]
-        boxes = torch.as_tensor(boxes).reshape(-1, 4)  # guard against no boxes
+        # guard against no boxes via resizing
+        boxes = torch.as_tensor(boxes, dtype=torch.float32).reshape(-1, 4)
         boxes[:, 2:] += boxes[:, :2]
         boxes[:, 0::2].clamp_(min=0, max=w)
         boxes[:, 1::2].clamp_(min=0, max=h)
 
         classes = [obj["category_id"] for obj in anno]
-        classes = torch.tensor(classes)
+        classes = torch.tensor(classes, dtype=torch.int64)
 
         segmentations = [obj["segmentation"] for obj in anno]
         masks = convert_coco_poly_to_mask(segmentations, h, w)
